Refactor main compiler class by splitting into three files
